pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chronos"
version = "1.0.0"
description = "CHRONOS - Continuous Heuristic Recognition of Network Operations & Signatures"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "CHRONOS Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "redis>=4.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "elasticsearch>=8.0.0",
    "neo4j>=5.0.0",
    "kafka-python>=2.0.2",
    "aiohttp>=3.8.0",
    "websockets>=11.0.0",
    "requests>=2.31.0",
    "cryptography>=39.0.0",
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
    "bcrypt>=4.0.0",
    "dnspython>=2.4.0",
    "python-dateutil>=2.8.2",
    "python-nmap>=0.7.1",
    "scikit-learn>=1.3.0",
    "plotly>=5.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/chronos-security/chronos"
Documentation = "https://docs.chronos.security"
Repository = "https://github.com/chronos-security/chronos"

[tool.setuptools.packages.find]
where = ["."]
include = ["chronos*"]

[tool.setuptools.package-data]
chronos = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
